adds initial drink build functions
Allows user to start the first drink, select a base using the various themed pages, see the values inputted from the preset base, and add a new drink using the same method as the original. New drinks are automatically selected and created when the user deselects the current drink and then picks a starting page. Also improves on state management and further centralizes lib information